+++ tests/Neoxygen/NeoClient/Tests/Schema/SchemaIntegrationTest.php
@@ -4,6 +4,7 @@ namespace Neoxygen\NeoClient\Tests\Schema;
 
 use Neoxygen\NeoClient\ClientBuilder;
 use Neoxygen\NeoClient\Schema\Index;
+use Neoxygen\NeoClient\Schema\UniqueConstraint;
 use Neoxygen\NeoClient\Tests\Helper\SchemaContainsIndexConstraint;
 
 class SchemaIntegrationTest extends GraphUnitTestCase
@@ -62,7 +63,51 @@ class SchemaIntegrationTest extends GraphUnitTestCase
         $this->client->dropSchemaIndex($index);
 
         $this->assertCount(0, $this->client->getSchemaIndexes());
+    }
 
+    /**
+     * @group schema
+     * @group integration
+     */
+    public function testUniqueConstraintIsCreated()
+    {
+        $constraint = $this->client->createSchemaUniqueConstraint("UniqueNode", "id");
+        $this->assertInstanceOf('Neoxygen\NeoClient\Schema\UniqueConstraint', $constraint);
+        $this->assertUniqueConstraintIsLive($constraint);
+
+    }
+
+    /**
+     * @group schema
+     * @group integration
+     */
+    public function testDropUniqueConstraint()
+    {
+        $constraint = $this->client->createSchemaUniqueConstraint("UniqueNode", "id");
+        $this->client->dropSchemaUniqueConstraint($constraint);
+
+        $this->assertCount(0, $this->client->getSchemaUniqueConstraints());
+    }
+
+    /**
+     * @group schema
+     * @group integration
+     */
+    public function testMultipleConstraintsAreReturned()
+    {
+        $this->client->createSchemaUniqueConstraint("UniqueNode", "id");
+        $this->client->createSchemaUniqueConstraint("UniqueNode", "login");
+        $this->assertCount(2, $this->client->getSchemaUniqueConstraints());
+    }
+
+    private function assertUniqueConstraintIsLive(UniqueConstraint $index)
+    {
+        $indexes = $this->client->getSchemaUniqueConstraints();
+
+        $this->assertThat(
+          $index,
+          new SchemaContainsIndexConstraint($indexes)
+        );
     }
 
     private function assertIndexIsLive(Index $index)