+++ tests/Neoxygen/NeoClient/Tests/Schema/SchemaUnitTest.php
@@ -3,6 +3,7 @@
 namespace Neoxygen\NeoClient\Tests\Schema;
 
 use Neoxygen\NeoClient\Schema\Index;
+use Neoxygen\NeoClient\Schema\UniqueConstraint;
 
 class SchemaUnitTest extends \PHPUnit_Framework_TestCase
 {
@@ -15,5 +16,15 @@ class SchemaUnitTest extends \PHPUnit_Framework_TestCase
         $index = new Index("User", "id");
         $this->assertEquals("User", $index->getLabel());
         $this->assertEquals("id", $index->getProperty());
+    }
+
+    /**
+     * @group schema
+     */
+    public function testSchemaConstraintIsCreatedWithLabelAndName()
+    {
+        $constraint = new UniqueConstraint("User", "id");
+        $this->assertEquals("User", $constraint->getLabel());
+        $this->assertEquals("id", $constraint->getProperty());
     }
 }